Blog list tests, seperate files for tests
